# src/main/resources/openapi/matawan-test-api.yaml
openapi: 3.0.3
info:
  title: Matawan Test API
  description: API de gestion de l'effectif de l'OGC Nice
  version: 1.0.0
paths:
  /api/teams:
    get:
      tags:
        - Teams
      operationId: getTeams
      summary: Récupère la liste paginée des équipes
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          schema:
            type: string
            default: name
          description: "Propriété de tri (name, acronym, budget)"
      responses:
        '200':
          description: Page d'équipes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageTeamResponse'
    post:
      tags:
        - Teams
      operationId: addTeam
      summary: Ajoute une nouvelle équipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamCreateRequest'
      responses:
        '201':
          description: Équipe créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    TeamCreateRequest:
      type: object
      required:
        - name
        - acronym
        - budget
      properties:
        name:
          type: string
          description: Nom de l'équipe
          example: OGC Nice
        acronym:
          type: string
          description: Acronyme de l'équipe
          example: OGCN
        budget:
          type: number
          format: double
          description: Budget en euros
          minimum: 0
          example: 10000000
        players:
          type: array
          description: Liste optionnelle des joueurs
          items:
            $ref: '#/components/schemas/PlayerRequest'
    PlayerRequest:
      type: object
      required:
        - name
        - position
      properties:
        name:
          type: string
          example: Dante
        position:
          type: string
          enum: [GOALKEEPER, DEFENDER, MIDFIELDER, FORWARD]
          example: DEFENDER
    TeamResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        acronym:
          type: string
        budget:
          type: number
        players:
          type: array
          items:
            $ref: '#/components/schemas/PlayerResponse'
    PlayerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        position:
          type: string
    PageTeamResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TeamResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: string